requirements.txt
requests>=2.25.1
fastapi>=0.68.0
uvicorn>=0.15.0
arq>=0.25.0
redis>=4.0.0
pydantic>=2.0.0
python-dotenv>=0.19.0
boto3>=1.18.0
beautifulsoup4>=4.9.3
tenacity>=9.1.3
rnet>=2.4.2
# Minimal set required for the FastAPI service architecture and scraper